use chia.exe to determine version and find config
